requirements.txt
streamlit
numpy
plotly
qdrant-edge-py